backend/app/policy_store.py
"""
Policy Store - Storage and Management of Policies
"""
from typing import List, Optional
from datetime import datetime
import os
from dotenv import load_dotenv

from .models import Policy, PolicyStatus

load_dotenv()


class PolicyStore:
    """
    Policy store for managing business policies.
    
    In production, this would connect to Neo4j or another database.
    Currently uses in-memory storage for development.
    """
    
    def __init__(self):
        self._policies: dict[str, Policy] = {}
        self._neo4j_uri = os.getenv("NEO4J_URI")
        self._neo4j_user = os.getenv("NEO4J_USER")
        self._neo4j_password = os.getenv("NEO4J_PASSWORD")
    
    async def create_policy(self, policy: Policy) -> Policy:
        """Create a new policy"""
        policy.id = f"policy_{len(self._policies) + 1}"
        policy.created_at = datetime.utcnow()
        policy.updated_at = datetime.utcnow()
        self._policies[policy.id] = policy
        return policy
    
    async def get_policy(self, policy_id: str) -> Optional[Policy]:
        """Get a policy by ID"""
        return self._policies.get(policy_id)
    
    async def get_all_policies(self) -> List[Policy]:
        """Get all policies"""
        return list(self._policies.values())
    
    async def get_active_policies(self) -> List[Policy]:
        """Get all active policies"""
        return [p for p in self._policies.values() if p.status == PolicyStatus.ACTIVE]
    
    async def update_policy(self, policy_id: str, updates: dict) -> Optional[Policy]:
        """Update an existing policy"""
        if policy_id not in self._policies:
            return None
        
        policy = self._policies[policy_id]
        for key, value in updates.items():
            if hasattr(policy, key):
                setattr(policy, key, value)
        policy.updated_at = datetime.utcnow()
        return policy
    
    async def delete_policy(self, policy_id: str) -> bool:
        """Delete a policy"""
        if policy_id in self._policies:
            del self._policies[policy_id]
            return True
        return False
    
    async def connect_neo4j(self):
        """
        Connect to Neo4j database.
        
        TODO: Implement actual Neo4j connection using neo4j-python-driver
        """
        # Placeholder for Neo4j connection
        pass


# Singleton instance
policy_store = PolicyStore()
