backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0

# Google Gmail API
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Neo4j
neo4j==5.17.0

# Anthropic/Gemini
google-generativeai==0.3.2

# Utilities
pydantic==2.6.0
httpx==0.26.0
